Move parathreads screen inside Parachains
